+++ R/adjust.R
@@ -71,7 +71,7 @@ adjust_infection_to_report <- function(infections, delay_defs,
                                        reporting_model, reporting_effect,
                                        type = "sample",
                                        truncate_future = TRUE) {
-  
+
   # Reset DT Defaults on Exit
   set_dt_single_thread()
 
@@ -131,7 +131,7 @@ adjust_infection_to_report <- function(infections, delay_defs,
   }
 
   ## Truncate reported cases by maximum infection date
-  if (type %in% "sample" & truncate_future) {
+  if (type %in% "sample" && truncate_future) {
     report <- report[date <= max(infections$date)]
   }
   return(report)