+++ R/epinow.R
@@ -7,9 +7,11 @@
 #' results and interpreting them. See [here](https://gist.github.com/seabbs/163d0f195892cde685c70473e1f5e867) for an
 #' example of using `epinow` to estimate Rt for Covid-19 in a country from the ECDC data source.
 #' @param output A character vector of optional output to return. Supported options are samples ("samples"),
-#' plots ("plots"), the run time ("timing"), copying the dated folder into a latest folder (if `target_folder` is not null,
-#' set using "latest"), and the stan fit ("fit"). The default is to return all options. This argument uses partial matching
-#' so for example passing "sam" will lead to samples being reported.
+#' plots ("plots"), the run time ("timing"), copying the dated folder into a
+#' latest folder (if `target_folder` is not null,
+#' set using "latest"), and the stan fit ("fit"). The default is to return all
+#' options. This argument uses partial matching so for example passing "sam"
+#' will lead to samples being reported.
 #' @param return_output Logical, defaults to FALSE. Should output be returned, this automatically updates to TRUE
 #' if no directory for saving is specified.
 #' @param plot_args A list of optional arguments passed to `plot.epinow()`.
@@ -55,7 +57,7 @@
 #'
 #' # summary of R estimates
 #' summary(out, type = "parameters", params = "R")
-#' 
+#'
 #' options(old_opts)
 #' }
 epinow <- function(reported_cases,